proxmox/requirements.txt
flask
flask-compress
requests
pyyaml
gunicorn